requirements.txt
numpy==1.26.4
scipy==1.12.0

opencv-contrib-python==4.10.0.84
mediapipe==0.10.21

sounddevice>=0.4.6
soundfile>=0.12

python_speech_features>=0.6
scenedetect==0.5.1

torch==2.2.2
torchvision==0.17.2

transformers>=4.35
huggingface-hub>=0.17
accelerate>=0.24
safetensors>=0.4

tqdm>=4.66
Pillow>=10.0

protobuf>=4.25.3,<5